:recycle: refactor: re-organize file&folder locate
