+++ src/auth/auth.module.ts
@@ -6,7 +6,7 @@ import { ConfigService } from '@nestjs/config';
 
 import { AuthController } from './auth.controller';
 import { AuthService } from './auth.service';
-import { UserEntity } from './user.entity';
+import { UserEntity } from './struct/user.entity';
 import { jwtConfig } from 'src/common/config/jwt.config';
 import { JwtStrategy } from './jwt.strategy';
 import { UserManager, UserReader } from './user.handler';
