--- src/auth/user.entity.ts
@@ -1,23 +1,0 @@
-import {
-  BaseEntity,
-  Column,
-  Entity,
-  PrimaryGeneratedColumn,
-  Unique,
-} from 'typeorm';
-
-@Entity()
-@Unique(['email'])
-export class UserEntity extends BaseEntity {
-  @PrimaryGeneratedColumn()
-  id: number;
-
-  @Column()
-  email: string;
-
-  @Column()
-  password: string;
-
-  @Column({ nullable: false, default: 0 })
-  balance: number;
-}
