+++ src/payment/payment.controller.ts
@@ -8,7 +8,7 @@ import {
 } from '@nestjs/common';
 import { PaymentService } from './payment.service';
 import { AuthGuard } from '@nestjs/passport';
-import { GetUser } from 'src/auth/get-user.decorator';
+import { GetUser } from 'src/common/decorator/get-user.decorator';
 import { UserEntity } from 'src/auth/struct/user.entity';
 
 @Controller('payment')
