+++ src/reservation/reservation.controller.ts
@@ -7,7 +7,7 @@ import {
 } from '@nestjs/common';
 import { AuthGuard } from '@nestjs/passport';
 import { ReservationService } from './reservation.service';
-import { GetUser } from 'src/auth/get-user.decorator';
+import { GetUser } from 'src/common/decorator/get-user.decorator';
 import { UserEntity } from 'src/auth/struct/user.entity';
 import { RequestReservationDto } from './dto/request-reservation.dto';
 import { ReservationEntity } from './reservation.entity';
